Refactor and increase test coverage.
